+++ search-films/src/api/film.request.ts
@@ -44,7 +44,12 @@ export type FilmDescApi = {
     '#TITLE': string, // name of the film
     '#IMDB_ID': string, // id of the film
     '#IMG_POSTER': string, // url of the film's poster 
+	rating?: number // rating of the film (optional)
 }
+
+export type FilmsDescApi = {
+	description: FilmDescApi[]
+};
 
 // !!!!! CAN THROW RUNTIME ERROR
 /**
@@ -54,12 +59,23 @@ export type FilmDescApi = {
  */
 export async function searchFilmDescByName(filmName: string) {
 	try {
-		const response = await axios.get(`${URL_TO_DATA}`, {
+		const filmDesc = await axios.get<FilmsDescApi>(`${URL_TO_DATA}`, {
 			params: {
 				[FILM_NAME_PARAMETER]: filmName
 			}
+		}); // FilmDescApi[] without rating
+		const ratingPromises = filmDesc.data
+			.description.map(async (description: FilmDescApi) => {
+				return getDetails(description['#IMDB_ID'])
+				;
+			});
+		
+		const ratings = await Promise.all(ratingPromises);
+		const res = filmDesc.data.description.map((description, index) => {
+			return {...description, rating: ratings[index].aggregateRating.ratingValue}
 		});
-		return response.data.description as FilmDescApi[];
+
+		return res;
 	} catch (err) {
 		console.error(err);
 		throw new Error('Failed to fetch film description');